+++ app/src/main/java/com/link_up/matching_manager/ui/my/team/MyTeamApplicationEditActivity.kt
@@ -448,18 +448,28 @@ class MyTeamApplicationEditActivity : AppCompatActivity() {
                 }
         }
         else {
-            binding.progressBar.visibility = View.VISIBLE
+            if (data.postImg == "") {
+                binding.progressBar.visibility = View.VISIBLE
+                viewModel.editApplication(data, newData)
+                binding.progressBar.visibility = View.INVISIBLE
+                Toast.makeText(this, "게시글이 수정되었습니다.", Toast.LENGTH_SHORT).show()
 
-            fileRef.delete()
-                .addOnSuccessListener {
-                    viewModel.editApplication(data, newData)
-                    binding.progressBar.visibility = View.INVISIBLE
-                    Toast.makeText(this, "게시글이 수정되었습니다.", Toast.LENGTH_SHORT).show()
-                }
-                .addOnFailureListener { exception ->
-                    binding.progressBar.visibility = View.INVISIBLE
-                    Toast.makeText(this, "게시글 수정을 실패하였습니다. 다시 시도해 주세요.", Toast.LENGTH_SHORT).show()
-                }
+            }
+            else {
+                binding.progressBar.visibility = View.VISIBLE
+
+                fileRef.delete()
+                    .addOnSuccessListener {
+                        viewModel.editApplication(data, newData)
+                        binding.progressBar.visibility = View.INVISIBLE
+                        Toast.makeText(this, "게시글이 수정되었습니다.", Toast.LENGTH_SHORT).show()
+                    }
+                    .addOnFailureListener { exception ->
+                        binding.progressBar.visibility = View.INVISIBLE
+                        Toast.makeText(this, "게시글 수정을 실패하였습니다. 다시 시도해 주세요.", Toast.LENGTH_SHORT).show()
+                    }
+            }
+
         }
     }
 }
